(Tech) refactor(AppButtons): remove duplicate code
