+++ src/ui/components/buttons/AppButton/AppButton.web.tsx
@@ -1,0 +1,106 @@
+import React, { CSSProperties, memo, MouseEventHandler, useCallback } from 'react'
+import styled from 'styled-components'
+
+import { accessibilityAndTestId } from 'libs/accessibilityAndTestId'
+import { AppButtonInner } from 'ui/components/buttons/AppButton/AppButtonInner'
+import { DefaultLoadingIndicator } from 'ui/components/buttons/AppButton/DefaultLoadingIndicator'
+import { buttonWebStyles } from 'ui/components/buttons/AppButton/styleUtils'
+import {
+  AppButtonEventWeb,
+  AppButtonProps,
+  OnlyBaseButtonProps,
+  TouchableOpacityButtonProps,
+} from 'ui/components/buttons/AppButton/types'
+
+const _AppButton = <T extends AppButtonProps>({
+  icon: Icon,
+  title: Title,
+  inline,
+  disabled,
+  isLoading,
+  onPress,
+  onLongPress,
+  mediumWidth,
+  fullWidth,
+  buttonHeight = 'small',
+  inlineHeight,
+  accessibilityLabel,
+  accessibilityDescribedBy,
+  testId,
+  wording,
+  adjustsFontSizeToFit,
+  justifyContent,
+  numberOfLines,
+  style,
+  loadingIndicator: LoadingIndicator,
+  center,
+  type = 'button',
+  className,
+  name,
+}: OnlyBaseButtonProps<T>) => {
+  const pressHandler = disabled || isLoading ? undefined : (onPress as AppButtonEventWeb)
+  const longPressHandler = disabled || isLoading ? undefined : (onLongPress as AppButtonEventWeb)
+
+  const onClick: MouseEventHandler<HTMLButtonElement> = useCallback(
+    (event) => {
+      if (type === 'submit' && pressHandler) {
+        event.preventDefault()
+      }
+      if (pressHandler) {
+        pressHandler(event)
+      }
+    },
+    [type, pressHandler]
+  )
+
+  const onDoubleClick: MouseEventHandler<HTMLButtonElement> = useCallback(
+    (event) => {
+      if (type === 'submit' && pressHandler) {
+        event.preventDefault()
+      }
+      if (longPressHandler) {
+        longPressHandler(event)
+      }
+    },
+    [type, longPressHandler]
+  )
+  return (
+    <Button
+      {...accessibilityAndTestId(accessibilityLabel, testId)}
+      name={name}
+      onClick={onClick}
+      onDoubleClick={onDoubleClick}
+      disabled={disabled}
+      type={type}
+      aria-describedby={accessibilityDescribedBy}
+      mediumWidth={mediumWidth}
+      fullWidth={fullWidth}
+      buttonHeight={buttonHeight}
+      inline={inline}
+      inlineHeight={inlineHeight}
+      justifyContent={justifyContent}
+      numberOfLines={numberOfLines}
+      style={style as CSSProperties}
+      center={center}
+      className={className}>
+      <AppButtonInner
+        loadingIndicator={LoadingIndicator}
+        isLoading={isLoading}
+        icon={Icon}
+        title={Title}
+        adjustsFontSizeToFit={adjustsFontSizeToFit}
+        numberOfLines={numberOfLines}
+        wording={wording}
+      />
+    </Button>
+  )
+}
+
+_AppButton.defaultProps = {
+  loadingIndicator: DefaultLoadingIndicator,
+}
+
+// memo is used to avoid useless rendering while props remain unchanged
+export const AppButton = memo(_AppButton)
+
+const Button = styled.button<TouchableOpacityButtonProps>(buttonWebStyles)
