+++ src/ui/components/buttons/ButtonPrimary.tsx
@@ -1,7 +1,8 @@
 import { Platform } from 'react-native'
 import styled from 'styled-components/native'
 
-import { AppButton, BaseButtonProps } from 'ui/components/buttons/AppButton'
+import { AppButton } from 'ui/components/buttons/AppButton/AppButton'
+import { BaseButtonProps } from 'ui/components/buttons/AppButton/types'
 import { styledButton } from 'ui/components/buttons/styledButton'
 import { Logo as InitialLoadingIndicator } from 'ui/svg/icons/Logo'
 import { Typo } from 'ui/theme'
@@ -42,15 +43,18 @@ export const ButtonPrimary = styledButton(AppButton).attrs<BaseButtonProps>(
     backgroundColor = theme.buttons.disabled.primary.backgroundColor
   }
 
+  let webOnly = {}
+  if (Platform.OS === 'web') {
+    webOnly = {
+      ['&:disabled']: {
+        backgroundColor: theme.buttons.disabled.primary.backgroundColor,
+      },
+    }
+  }
+
   return {
     backgroundColor,
-    ...(Platform.OS === 'web'
-      ? {
-          ['&:disabled']: {
-            backgroundColor: theme.buttons.disabled.primary.backgroundColor,
-          },
-        }
-      : {}),
+    ...webOnly,
   }
 })
