+++ src/ui/components/buttons/ButtonPrimaryWhite.tsx
@@ -1,7 +1,8 @@
 import { Platform } from 'react-native'
 import styled from 'styled-components/native'
 
-import { AppButton, BaseButtonProps } from 'ui/components/buttons/AppButton'
+import { AppButton } from 'ui/components/buttons/AppButton/AppButton'
+import { BaseButtonProps } from 'ui/components/buttons/AppButton/types'
 import { styledButton } from 'ui/components/buttons/styledButton'
 import { Logo as InitialLoadingIndicator } from 'ui/svg/icons/Logo'
 import { Typo } from 'ui/theme'
@@ -37,16 +38,21 @@ export const ButtonPrimaryWhite = styledButton(AppButton).attrs<BaseButtonProps>
       title: Title,
     }
   }
-)(({ theme }) => ({
-  backgroundColor: theme.buttons.primaryWhite.backgroundColor,
-  ...(Platform.OS === 'web'
-    ? {
-        ['&:disabled']: {
-          backgroundColor: theme.buttons.disabled.primary.backgroundColor,
-        },
-      }
-    : {}),
-}))
+)(({ theme }) => {
+  let webOnly = {}
+  if (Platform.OS === 'web') {
+    webOnly = {
+      ['&:disabled']: {
+        backgroundColor: theme.buttons.disabled.primary.backgroundColor,
+      },
+    }
+  }
+
+  return {
+    backgroundColor: theme.buttons.primaryWhite.backgroundColor,
+    ...webOnly,
+  }
+})
 
 const LoadingIndicator = styled(InitialLoadingIndicator).attrs(({ theme }) => ({
   color: theme.buttons.primaryWhite.loadingIconColor,
