+++ src/ui/components/buttons/ButtonSecondary.tsx
@@ -1,7 +1,8 @@
 import { Platform } from 'react-native'
 import styled from 'styled-components/native'
 
-import { AppButton, BaseButtonProps } from 'ui/components/buttons/AppButton'
+import { AppButton } from 'ui/components/buttons/AppButton/AppButton'
+import { BaseButtonProps } from 'ui/components/buttons/AppButton/types'
 import { styledButton } from 'ui/components/buttons/styledButton'
 import { Logo as InitialLoadingIndicator } from 'ui/svg/icons/Logo'
 import { Typo } from 'ui/theme'
@@ -48,17 +49,20 @@ export const ButtonSecondary = styledButton(AppButton).attrs<BaseButtonProps>(
     borderColor = theme.buttons.disabled.secondary.borderColor
   }
 
+  let webOnly = {}
+  if (Platform.OS === 'web') {
+    webOnly = {
+      ['&:disabled']: {
+        borderColor: theme.buttons.disabled.secondary.borderColor,
+      },
+    }
+  }
+
   return {
     backgroundColor,
     borderColor,
     borderWidth,
-    ...(Platform.OS === 'web'
-      ? {
-          ['&:disabled']: {
-            borderColor: theme.buttons.disabled.secondary.borderColor,
-          },
-        }
-      : {}),
+    ...webOnly,
   }
 })
